Use a swipeable tab layout for Featured & Most Popular, instead of the Spinner dropdown. Removed youtube api key, instead referring to it from environment's gradle.properties file.
