+++ app/src/main/java/free/rm/skytube/gui/activities/MainActivity.java
@@ -17,6 +17,7 @@
 
 package free.rm.skytube.gui.activities;
 
+import android.app.ProgressDialog;
 import android.content.ClipDescription;
 import android.content.ClipboardManager;
 import android.content.DialogInterface;
@@ -29,49 +30,67 @@ import android.support.v4.view.MenuItemCompat;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.SearchView;
+import android.util.Log;
+import android.util.Pair;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.EditText;
 import android.widget.FrameLayout;
+import android.widget.Toast;
+
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.OutputStream;
+import java.net.URL;
 
 import butterknife.Bind;
 import butterknife.ButterKnife;
 import free.rm.skytube.R;
+import free.rm.skytube.businessobjects.AsyncTaskParallel;
 import free.rm.skytube.businessobjects.MainActivityListener;
 import free.rm.skytube.businessobjects.YouTubeChannel;
+import free.rm.skytube.gui.businessobjects.UpdatesChecker;
+import free.rm.skytube.gui.businessobjects.WebStream;
 import free.rm.skytube.gui.fragments.ChannelBrowserFragment;
+import free.rm.skytube.gui.fragments.MainFragment;
 import free.rm.skytube.gui.fragments.SearchVideoGridFragment;
-import free.rm.skytube.gui.fragments.VideosGridFragment;
 
 /**
  * Main activity (launcher).  This activity holds {@link free.rm.skytube.gui.fragments.VideosGridFragment}.
  */
 public class MainActivity extends AppCompatActivity implements MainActivityListener {
 	public static final String ACTION_VIEW_CHANNEL = "MainActivity.ViewChannel";
-	public static final String FEATURED_FRAGMENT = "MainActivity.FeaturedFragment";
+	public static final String MAIN_FRAGMENT = "MainActivity.MainFragment";
 	public static final String CHANNEL_BROWSER_FRAGMENT = "MainActivity.ChannelBrowserFragment";
 	public static final String SEARCH_FRAGMENT = "MainActivity.SearchFragment";
 
-
+	/** Set to true of the UpdatesCheckerTask has run; false otherwise. */
+	private static boolean updatesCheckerTaskRan = false;
 
 	@Bind(R.id.fragment_container)
 	FrameLayout fragmentContainer;
 
-	VideosGridFragment featuredFragment;
+	MainFragment mainFragment;
 	ChannelBrowserFragment channelBrowserFragment;
 	SearchVideoGridFragment searchVideoGridFragment;
 
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
+
+		// check for updates (one time only)
+		if (!updatesCheckerTaskRan)
+			new UpdatesCheckerTask().executeInParallel();
+
 		setContentView(R.layout.activity_main);
 		ButterKnife.bind(this);
 
 		if(fragmentContainer != null) {
 			if(savedInstanceState != null) {
-				featuredFragment = (VideosGridFragment)getSupportFragmentManager().getFragment(savedInstanceState, FEATURED_FRAGMENT);
+				mainFragment = (MainFragment)getSupportFragmentManager().getFragment(savedInstanceState, MAIN_FRAGMENT);
 				channelBrowserFragment = (ChannelBrowserFragment) getSupportFragmentManager().getFragment(savedInstanceState, CHANNEL_BROWSER_FRAGMENT);
 				searchVideoGridFragment = (SearchVideoGridFragment) getSupportFragmentManager().getFragment(savedInstanceState, SEARCH_FRAGMENT);
 			}
@@ -80,10 +99,9 @@ public class MainActivity extends AppCompatActivity implements MainActivityListe
 				YouTubeChannel channel = (YouTubeChannel) getIntent().getSerializableExtra(ChannelBrowserFragment.CHANNEL_OBJ);
 				onChannelClick(channel);
 			} else {
-				if(featuredFragment == null) {
-					featuredFragment = new VideosGridFragment();
-					featuredFragment.setArguments(getIntent().getExtras());
-					getSupportFragmentManager().beginTransaction().add(R.id.fragment_container, featuredFragment).commit();
+				if(mainFragment == null) {
+					mainFragment = new MainFragment();
+					getSupportFragmentManager().beginTransaction().add(R.id.fragment_container, mainFragment).commit();
 				}
 			}
 		}
@@ -91,8 +109,8 @@ public class MainActivity extends AppCompatActivity implements MainActivityListe
 
 	@Override
 	protected void onSaveInstanceState(Bundle outState) {
-		if(featuredFragment != null)
-			getSupportFragmentManager().putFragment(outState, FEATURED_FRAGMENT, featuredFragment);
+		if(mainFragment != null)
+			getSupportFragmentManager().putFragment(outState, MAIN_FRAGMENT, mainFragment);
 		if(channelBrowserFragment != null && channelBrowserFragment.isVisible())
 			getSupportFragmentManager().putFragment(outState, CHANNEL_BROWSER_FRAGMENT, channelBrowserFragment);
 		if(searchVideoGridFragment != null && searchVideoGridFragment.isVisible())
@@ -147,7 +165,7 @@ public class MainActivity extends AppCompatActivity implements MainActivityListe
 				displayEnterVideoUrlDialog();
 				return true;
 			case android.R.id.home:
-				if(featuredFragment == null || !featuredFragment.isVisible()) {
+				if(mainFragment == null || !mainFragment.isVisible()) {
 					onBackPressed();
 					return true;
 				}
@@ -243,5 +261,162 @@ public class MainActivity extends AppCompatActivity implements MainActivityListe
 		channelBrowserFragment.setArguments(args);
 
 		switchToFragment(channelBrowserFragment);
+	}
+
+	////////////////////////////////////////////////////////////////////////////////////////////////
+
+
+	private class UpdatesCheckerTask extends AsyncTaskParallel<Void, Void, UpdatesChecker> {
+
+		@Override
+		protected UpdatesChecker doInBackground(Void... params) {
+			UpdatesChecker updatesChecker = new UpdatesChecker();
+			updatesChecker.checkForUpdates();
+			return updatesChecker;
+		}
+
+		@Override
+		protected void onPostExecute(final UpdatesChecker updatesChecker) {
+			updatesCheckerTaskRan = true;
+
+			if (updatesChecker != null && updatesChecker.getLatestApkUrl() != null) {
+				new AlertDialog.Builder(MainActivity.this)
+								.setTitle(R.string.update_available)
+								.setMessage( String.format(getResources().getString(R.string.update_dialog_msg), updatesChecker.getLatestApkVersion()) )
+								.setPositiveButton(R.string.update, new DialogInterface.OnClickListener() {
+									@Override
+									public void onClick(DialogInterface dialog, int which) {
+										new UpgradeAppTask(updatesChecker.getLatestApkUrl()).executeInParallel();
+									}
+								})
+								.setNegativeButton(R.string.later, null)
+								.show();
+			}
+		}
+
+	}
+
+
+	/**
+	 * This task will download the remote APK file and it will install it for the user (provided that
+	 * the user accepts such installation).
+	 */
+	private class UpgradeAppTask extends AsyncTaskParallel<Void, Integer, Pair<File, Throwable>> {
+
+		private URL apkUrl;
+		private ProgressDialog downloadDialog;
+
+		private final String TAG = UpgradeAppTask.class.getSimpleName();
+
+
+		public UpgradeAppTask(URL apkUrl) {
+			this.apkUrl = apkUrl;
+		}
+
+
+		@Override
+		protected void onPreExecute() {
+			// setup the download dialog and display it
+			downloadDialog = new ProgressDialog(MainActivity.this);
+			downloadDialog.setMessage(getString(R.string.downloading));
+			downloadDialog.setProgressStyle(ProgressDialog.STYLE_HORIZONTAL);
+			downloadDialog.setProgress(0);
+			downloadDialog.setMax(100);
+			downloadDialog.setCancelable(false);
+			downloadDialog.setProgressNumberFormat(null);
+			downloadDialog.show();
+		}
+
+		@Override
+		protected Pair<File, Throwable> doInBackground(Void... params) {
+			File		apkFile;
+			Throwable	exception = null;
+
+			// try to download the remote APK file
+			try {
+				apkFile = downloadApk();
+			} catch (Throwable e) {
+				apkFile = null;
+				exception = e;
+			}
+
+			return new Pair<>(apkFile, exception);
+		}
+
+
+		/**
+		 * Download the remote APK file and return an instance of {@link File}.
+		 *
+		 * @return	A {@link File} instance of the downloaded APK.
+		 * @throws IOException
+		 */
+		private File downloadApk() throws IOException {
+			WebStream webStream = new WebStream(this.apkUrl);
+			File			apkFile = File.createTempFile("skytube-upgrade", ".apk", getCacheDir());
+			OutputStream out;
+
+			// set the APK file to readable to every user so that this file can be read by Android's
+			// package manager program
+			apkFile.setReadable(true /*set file to readable*/, false /*set readable to every user on the system*/);
+			out = new FileOutputStream(apkFile);
+
+			// download the file by transferring bytes from in to out
+			byte[]	buf = new byte[1024];
+			int		totalBytesRead = 0;
+			for (int bytesRead; (bytesRead = webStream.getStream().read(buf)) > 0; ) {
+				out.write(buf, 0, bytesRead);
+
+				// update the progressbar of the downloadDialog
+				totalBytesRead += bytesRead;
+				publishProgress(totalBytesRead, webStream.getStreamSize());
+			}
+
+			// close the streams
+			webStream.getStream().close();
+			out.close();
+
+			return apkFile;
+		}
+
+
+		@Override
+		protected void onProgressUpdate(Integer... values) {
+			float	totalBytesRead = values[0];
+			float	fileSize = values[1];
+			float	percentageDownloaded = (totalBytesRead / fileSize) * 100f;
+
+			downloadDialog.setProgress((int)percentageDownloaded);
+		}
+
+
+		@Override
+		protected void onPostExecute(Pair<File, Throwable> out) {
+			File		apkFile   = out.first;
+			Throwable	exception = out.second;
+
+			// hide the download dialog
+			downloadDialog.dismiss();
+
+			if (exception != null) {
+				Log.e(TAG, "Unable to upgrade app", exception);
+				Toast.makeText(MainActivity.this, R.string.update_failure, Toast.LENGTH_LONG).show();
+			} else {
+				displayUpgradeAppDialog(apkFile);
+			}
+		}
+
+
+		/**
+		 * Ask the user whether he wants to install the latest SkyTube's APK file.
+		 *
+		 * @param apkFile	APK file to install.
+		 */
+		private void displayUpgradeAppDialog(File apkFile) {
+			Intent intent = new Intent(Intent.ACTION_VIEW);
+			intent.setDataAndType(Uri.fromFile(apkFile), "application/vnd.android.package-archive");
+			intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);	// asks the user to open the newly updated app
+			startActivity(intent);
+		}
+
 	}
 }
