--- app/src/main/res/layout/activity_channel_browser.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<fragment xmlns:android="http://schemas.android.com/apk/res/android"
-		  android:id="@+id/channel_fragment"
-		  android:name="free.rm.skytube.gui.fragments.ChannelBrowserFragment"
-		  android:layout_width="match_parent"
-		  android:layout_height="match_parent" />
