--- app/src/main/res/layout/fragment_videos_grid.xml
@@ -1,27 +1,0 @@
-<android.support.v4.widget.DrawerLayout
-	xmlns:android="http://schemas.android.com/apk/res/android"
-	android:id="@+id/subs_drawer_layout"
-	android:layout_width="match_parent"
-	android:layout_height="match_parent">
-
-	<LinearLayout
-		android:layout_width="match_parent"
-		android:layout_height="match_parent"
-		android:orientation="vertical">
-		<include
-			android:id="@+id/activity_main_toolbar"
-			layout="@layout/activity_main_toolbar"/>
-
-		<include
-			android:layout_width="match_parent"
-			android:layout_height="match_parent"
-			layout="@layout/videos_gridview"/>
-	</LinearLayout>
-
-	<include
-		android:layout_width="240dp"
-		android:layout_height="match_parent"
-		android:layout_gravity="start"
-		layout="@layout/subs_drawer"/>
-
-</android.support.v4.widget.DrawerLayout>
