--- app/src/main/res/values/strings_api_key.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-	<string name="API_KEY">AIzaSyDAOcAgAAd098p_FLIePaQNKPTpcyhUGBA</string>
-</resources>
