+++ app/src/nonFree/java/free/rm/skytube/gui/businessobjects/YouTubePlayer.java
@@ -25,6 +25,7 @@ import android.content.SharedPreferences;
 import android.preference.PreferenceManager;
 import android.util.Log;
 
+import free.rm.skytube.BuildConfig;
 import free.rm.skytube.R;
 import free.rm.skytube.businessobjects.YouTubeVideo;
 import free.rm.skytube.gui.activities.YouTubePlayerActivity;
@@ -72,7 +73,7 @@ public class YouTubePlayer {
 	private static void launchOfficialYouTubePlayer(String videoId, Context context) {
 		try {
 			// try to start the YouTube standalone player
-			Intent intent = com.google.android.youtube.player.YouTubeStandalonePlayer.createVideoIntent((Activity) context, context.getString(R.string.API_KEY), videoId);
+			Intent intent = com.google.android.youtube.player.YouTubeStandalonePlayer.createVideoIntent((Activity) context, BuildConfig.YOUTUBE_API_KEY, videoId);
 			context.startActivity(intent);
 		} catch (Exception e) {
 			String errorMsg = context.getString(R.string.launch_offical_player_error);
